refactorizando crud's
