+++ src/main/java/com/example/stose/StoseApplication.java
@@ -4,11 +4,19 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
 
-import com.example.stose.entidad.Libro;
-import com.example.stose.repositorio.LibroRepositorio;
+import com.example.stose.entity.Libro;
+import com.example.stose.repositories.LibroRepositorio;
+
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @SpringBootApplication
+@EnableSwagger2
 public class StoseApplication implements CommandLineRunner{
 
 	public static void main(String[] args) {
@@ -22,6 +30,15 @@ public class StoseApplication implements CommandLineRunner{
 	public void run(String... args) throws Exception {
 	 
 
+	}
+
+	@Bean
+	  public Docket blogApi() {
+	    return new Docket(DocumentationType.SWAGGER_2)
+	        .select()
+	          .apis(RequestHandlerSelectors.any())
+	          .paths(PathSelectors.any())
+	          .build();
 	}
 
 }
