+++ src/main/java/com/example/stose/repositories/IBlogDAO.java
@@ -3,7 +3,7 @@ package com.example.stose.repositories;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.stose.models.BlogDTO;
+import com.example.stose.model.BlogDTO;
 
 @Repository
 public interface IBlogDAO extends MongoRepository<BlogDTO, String> {
