+++ src/main/java/com/example/stose/repositories/LibroRepositorie.java
@@ -1,9 +1,9 @@
-package com.example.stose.repositorio;
+package com.example.stose.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.stose.entidad.Libro;
+import com.example.stose.entity.Libro;
 
 @Repository
 public interface LibroRepositorio extends JpaRepository<Libro, Long>{
